--- modules/new_serverpod_auth/serverpod_auth_jwt/serverpod_auth_jwt_server/lib/src/util/random_bytes.dart
@@ -1,11 +1,0 @@
-import 'dart:math';
-import 'dart:typed_data';
-
-final Random _random = Random.secure();
-
-/// Generates a list of secure random bytes of the specified length.
-Uint8List generateRandomBytes(final int length) {
-  return Uint8List.fromList(
-    List<int>.generate(length, (final int i) => _random.nextInt(256)),
-  );
-}
